UI: Data Table, instantiate DataTables in JS
